Simplified the query chaining functions.
Functions can not be added in the chain anymore and all collected
cursors are passed to the callback.

I was not able to think of use-cases where it was needed to add
functions to a chain that intercept cursors and pass their results to
other queries. This also gets to complex and leads to instability.