Removed usage of tornado.gen and moved execute/callproc/mogrify.
The requests per second jumped from ~1600 to ~1900 on my laptop
after the usage of tornado.gen has been reduced to zero. This
adds a bit more complexity and code duplication to the code, but
it's worth it considering the hude performance improvement.

While refactoring the connection pool I also moved the execute,
callproc and mogrify methods to the `Connection` class in
preparation for transaction support.